Spanish translation
Here I add Spanish translations for the labels .yml files, as well as the config\locales\es.yml file.
I have also created a Spanish version of the \admin\sources\_template_help erb in case it's possible to include it as part of the Spanish version of the platform. I noticed that in other languages this part of Muscat remained in English.