Require default EDITOR_PROFILE to be set as "default"
Most configuration values in config/application.rb (as copied from
config/sample_application.rb) have an explicit "default" value, except
for EDITOR_PROFILE, that may have an empty string. This complicates
the logic when this variable is used in Muscat, as it has to be
compared to empty string and then set to "default" before using it.