Reorder catalogue.rb code like source.rb
Catalogue record type shares with Source that both are bibliographic Marc
records.  In Muscat, Source is more complete and has more options, like the
ability to choose different templates.  This patch does not add or modifies
anything at all.  It just reorders the lines of code of catalogue.rb so that
the common ones are aligned with source.db and is easier to spot the
differences.