Allow different templates for secondary literature
This patch is the second version of the first step towards having different
templates for secondary literature.  It synchronises some code and
configuration files from sources so now, when creating a new record, appears
a menu with a single choice, the legacy template.

This version takes into account the catalogues to publications rename and it
adds the locale entries compatible with the i18n reorganisation.

There are yet no new fields (like record type) or tag changes.