New global config variable: MAIN_BIBLIOGRAPHIC_RECORD_TYPE
Muscat was created for cataloguing musical manuscripts (Sources), but it
also allows cataloguing secondary literature (Publications).  There is some
work going on to adapt Muscat to be used as institutional or research
repository, where the main bibliographic record type would be Publications.
This first patch allows to define this parameter, and just corrects the
public, Blacklight Solr interface.