Rename duplicated unique_records index
Two tables have the same "unique_records" index name.  Although this seems
not to be a problem with current MySQL/MariaDB versions, it may be a problem
in the future, or using other database softwares, like SQLite (not
supported, I know, but helped me to detect this duplication).

It is safer to have unique index name.  This migration renames both.