Normalise full_name_d removing diacritics and final punctuation
Documentation for Person model states that full_name_d should be normalised
to downcase and with diacritics removed, but currently it isn't so.  This
patch uses Rails I18n.transliterate, with the known limitation that it only
works for Latin alphabets.  As for other alphabets it returns a string full
of question marks, this patch reverses this damage using the current simpler
downcase normalisation.  Fixes #1165.