Add a to_h method to Marc tags to access its subfields
Allow Marc subfields to be accessed like a hash, so it is easy to extract
all subfields of a tag or check the existence or not of a subfield.
Optionally, values can be returned as a list (as_list boolean parameter, for
repetable subfields), and final punctuation can also be optionally stripped
(strip_punctuation boolean parameter).