Implement $a and $b in institutions
Institutional authority records and indices have the added complexity that
the name is split in two subfields: $a for corporate name and $b for
subordinate unit (see https://www.loc.gov/marc/authority/ad110.html).

This patch addresses both the authority records attributes and indices in
Muscat with the following principles in mind:

* Keep the changes as minimal as possible.
* Do not change Muscat current behaviour if no $b is used in any
  institutional field.

The patch keeps the current name attribute, but now it is the concatenation
of $a and $b (not just $a as before), and adds a new one just for $a (called
corporate_name), and another for just $b (called subordinate_unit), that
need a migration.

This patch adjustes yaml config files for marc institution, person and
source.

Closes #1196 following the discussion at #1171