Implement $a and $b in institutions, v2
Institutional authority records and indices have the added complexity that
the name is split in two subfields: $a for corporate name and $b for
subordinate unit (see https://www.loc.gov/marc/authority/ad110.html).

Following @xhero request in #1287, this second version renames the old
`name` attribute to `full_name`, while keeping `corporate_name` for $a and
`subordinate_unit` for $b.  This attribute rename forces changes in other
files, so it is larger and more invasive than the first one

This version has two limitations or drawbacks:

While testing it, I kept finding and "undefined method `name' for
many hours of unsuccessful debugging, the only cure was an "alias_attribute
:name, :full_name", which, I know, only hides the problem.

The second one is that autocomplete for institutions fields does no longer
work, and I haven't been able to find why.