Implement $a and $b in institutions, v3
Institutional authority records and indices have the added complexity that
the name is split in two subfields: $a for corporate name and $b for
subordinate unit (see https://www.loc.gov/marc/authority/ad110.html).

Following @xhero request in #1287, this second version renames the old name
attribute to full_name, while keeping corporate_name for $a and
subordinate_unit for $b.  This attribute rename forces changes in other
files, so it is larger and more invasive than the first one.

Regarding v2, this version:

* no longer needs an alias_attribute :name, :full_name
* autocomplete now works for institution fields
* no longer modifies Muscat default editing templates

I have been testing it and I hope this time is correct.