Allow items to be added and removed from folders
Currently the only way to add items to folders is via a batch selection form
a list created in the list view.  This patch allows to add and remove them
from the full record view.  Closes #931.

This second version unifies in a single (and hopefully complete) commit and
replaces #1388.

Thanks to @ivan-mr from @CodiTramuntana for his help finding a way of doing
it following the Rails conventions.