Move current filter information at top of sidebar
Muscat remembers search filters until they are explicity cleared, and that
confuses users because the list of records doesn't match with what is
expected.  Move the the information box to the top of the sidebar so it is
obvious that there are filters active.

Thanks to @ivan-mr from @CodiTramuntana for finding a css-only solution.

Closes #1251