Create links to display Marc records in Marc and Marcxml
Although Muscat catalog records in Marc, Marc formats are not easily
available.  Marcxml is accessible adding .xml at the end of any record, but
it is a format to exchange records between systems, not to be read by
catalogers.  This patch adds three links to Marc records: one to the default
Muscat view, another to a a simple, human readable but complete Marc21
record with all the fields and subfields, and a third on to the Marcxml one.

Maybe it closes #595