Make logs less verbose on development mode
It is difficult to find useful information in Muscat log files due to the
the noise that the workers write.  Some time ago I googled around to find an
easy to implement solution to this situation, as it should be common to any
Rails app.  I found this one to be simple, and works very well, at least for
me.