Pregenerate and use a person display_name to disambiguate similar ones
When different people share the same or similar name, other fields are used
to distinguish them, like life dates.  In Muscat, this richer form of name
is used, at least, in autocomplete forms, although it could be used on other
places.  Instead of building it each time, create it once and store it in db.