Add a clear button at top when a filter is active
When a filter is active, current value is shown at the top of the sidebar
(since the #1448 PR).  However, the button to clear the filter is at the
bottom.  For some records types with lots of filtering options, this can be
quite annoying as the clear button is down under, it forces to scroll
down a lot and distracts from the curren task.

This patch adds another clear button at top when there is an active filter,
to clear it out easily. Complements #1251.

Thanks to @ivan-mr from @CodiTramuntana for the implementation.

Maybe the muscat.scss patch needs some adjustments, as we were working on
our own personalized scss for each our sites, and the original styling seems
to inherit from upstream activeadmin gem.